+++ .mocharc.js
@@ -1,0 +1,3 @@
+module.exports = {
+    spec: "build/javascript/test/**/*.spec.js"
+};