+++ src/pdfjs/zoom/index.ts
@@ -1,0 +1,3 @@
+export { PinchZoomingInteraction } from "./pinch-zoom";
+export { DoubleTapZoomingInteractionImpl } from "./double-tap-zoom";
+export { DefaultZoomSettings } from "./settings";